fix(tui): gpu_free_bytes returns usize, matching the backend
Typo'd the accessor's return type as u64 in the previous commit; the CUDA
backend returns Option<usize>. My clippy run and the push raced in one command
line, so the broken commit reached the branch before its own gate had spoken —
run the gate, read it, then push.

+++ crates/spark-server/src/tui/data/mod.rs
@@ -14,7 +14,7 @@ pub mod metrics_poll;
 /// breaks the metal/CPU builds — which is exactly how the dashboard shipped:
 /// two call sites, no cfg, and the macOS CI job could not compile the crate.
 /// One gated accessor means the next caller cannot repeat that.
-pub fn gpu_free_bytes() -> Option<u64> {
+pub fn gpu_free_bytes() -> Option<usize> {
     #[cfg(feature = "cuda")]
     {
         spark_runtime::cuda_backend::cuda_free_memory_bytes()
